requirements.txt
torch>=2.1
torchvision>=0.16
pandas>=2.0
opencv-python>=4.8
Pillow>=10.0
numpy>=1.25
tqdm>=4.66
requests>=2.31.0
openpyxl>=3.1.0

pip install torch torchvision torchaudio
python -m streamlit run gui.py